Two commits that take user space genuinely 64-bit-shaped for the
regions the kernel owns end-to-end:

  Step 3 (3f30b286): user stack and sigreturn trampoline.
    USER_STACK_TOP        0xC000_0000 -> 0x0000_7FFF_FFFF_F000
    SIGRETURN_TRAMPOLINE  0xC000_0000 -> 0x0000_7FFE_FFFF_F000
    The stack + trampoline now live near the top of the canonical-low
    half (Linux x86_64-style). Page-table walks naturally handle the
    new high addresses; no user binary needs to be relinked because
    the stack address isn't baked into binaries.

  Step 4 (05954942): SYS_MMAP / SYS_MUNMAP get a u64 ABI on SYSCALL.
    The dispatcher now special-cases SYS_MMAP / SYS_MUNMAP alongside
    SYS_MMAP64 / SYS_MUNMAP64 and routes them to new u64 entry points
    (sys_mmap_u64 / sys_munmap_u64) that use the high-VA allocator
    (MMAP64_BASE 0x0000_0001_0000_0000 .. MMAP64_LIMIT 0x4000_0000_0000).
    libsyscall::mmap and stdlib::process::mmap propagate the full
    u64 address with u64::MAX as failure marker; libstdlib's heap
    sanity check follows.

After this merge:
  user stack            : upper half ✓
  sigreturn trampoline  : upper half ✓
  mmap-allocated pages  : upper half ✓
  programs (.text/.data): low half (0x0800_0000) — unchanged
  DLIBs                 : low half (0x0400_0000+) — unchanged
  brk / heap            : low half — unchanged

Programs and DLIBs still link at low-half fixed addresses; relinking
them requires updating six linker scripts plus consumers like
libs/uisys_client::UISYS_BASE and is tracked separately as Step 5
(estimated 6-12 days, deferred).

Verified for both commits: build green, image boots in QEMU/KVM
through compositor spawn, no exceptions.

+++ libs/libsyscall/src/lib.rs
@@ -350,9 +350,13 @@ pub fn sbrk(increment: u32) -> u64 {
 }
 
 /// Map anonymous pages. Returns address or `u64::MAX` on failure.
+///
+/// SYS_MMAP now uses the u64 ABI on the SYSCALL fast path: the kernel
+/// returns the full virtual address (in the upper canonical-low half,
+/// well above 4 GiB) or u64::MAX on error.
 pub fn mmap(size: u32) -> u64 {
     let ret = syscall1(SYS_MMAP, size as u64);
-    if ret == u32::MAX as u64 { u64::MAX } else { ret }
+    if ret == u64::MAX { u64::MAX } else { ret }
 }
 
 /// Unmap pages previously mapped with `mmap`. Returns 0 on success.
